shared/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agenthub-shared"
version = "0.1.0"
description = "Shared schemas, adapters, and parsers for Skills Registry Platform"
requires-python = ">=3.11"
dependencies = [
    "pydantic[email]>=2.0",
    "pyyaml>=6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["skills_registry_shared"]
